scripts/requirements.txt
gspread==5.11.0
google-auth==2.24.0
